Ingame Console, Text Rendering/Font Loading, Animations, and NPC Death
Press tilde (~) to toggle console

valid commands:
	die/close/quit - Exits game
	wireframe/wf - toggle wireframe/wf
	tfc - toggles free camera
	cs [sceneno] - changes scene to sceneno (if it exists)

Console logic written in LUA, so adding commands should be easy

The current implmentation of the AI is pretty weak, without collision awarness they quickly get stuck on stuff.

Bug in animation:
	I parse out the list of animations from the MD2 file and aggragate them based on name;
	unfortinalty, multiple animations can have the same name.
	So some animations loop go through the cycles of their neighbouring animations before ending.
	This can be seen on NPC death.

	Easy fix, but its there for the moment.

+++ src/Terrain.cpp
@@ -30,7 +30,7 @@ void terrain::genLightmap() {
 
 	int dirZ = -1;
 
-	float softness = 0.2;
+	float softness = 0.2f;
 
 	float minBrightness = 0.3f;
 	float maxBrightness = 8.0f;
@@ -42,7 +42,7 @@ void terrain::genLightmap() {
 		for (unsigned x = 0; x < dimensions; x++) {
 			int x1 = x - dirX;
 			int z1 = z - dirZ;
-			if (x1<0 || x1 >= dimensions || z1 <0 || z1 >= dimensions) {
+			if (x1<0 || x1 >= (int) dimensions || z1 <0 || z1 >= (int)dimensions) {
 				shade = 1.0f;
 			}
 			else {
